+++ cnipe.py
@@ -14,6 +14,7 @@ from PIL import Image, ImageOps
 from src.screenshot_taker import MyWidget
 from src.html_to_image import make_gradient_image
 from src.image_shaper import ImgModifier
+from src.scroll_action import ScrollableLabelButtonFrame
 
 customtkinter.set_appearance_mode("System")  # Modes: system (default), light, dark
 customtkinter.set_default_color_theme("green")  # Theme
@@ -50,6 +51,8 @@ class App(customtkinter.CTk):
         self.main_image = None
         self.modified_image = None
         self.padding = 25
+        self.scrolls = None
+        self.selected_grad_image = None
 
         # set grid layout 1x2
         self.grid_rowconfigure(0, weight=1)
@@ -85,7 +88,7 @@ class App(customtkinter.CTk):
 
         self.home_button_event()
 
-    def capture_screenshot(self, use_old=False, padding=25):
+    def capture_screenshot(self, use_old=False, new_gradient=None, padding=25):
         """
         capture screenshot
         """
@@ -103,19 +106,22 @@ class App(customtkinter.CTk):
             qt_app.exec_()
             window.destroy()
 
-        color_grad = None
-        if self.use_same_colors is not None and self.use_same_colors.get() == 1:
-            color_grad = self.colors
+        if new_gradient is None:
+            color_grad = None
+            if self.use_same_colors is not None and self.use_same_colors.get() == 1:
+                color_grad = self.colors
 
-        # create gradient
-        gradient_path, self.colors = make_gradient_image(
-            store_dir=IMAGE_DIR_PATH,
-            gradient_name=gradient_name,
-            colors=color_grad
-        )
+            # create gradient
+            self.selected_grad_image, self.colors = make_gradient_image(
+                store_dir=IMAGE_DIR_PATH,
+                gradient_name=gradient_name,
+                colors=color_grad
+            )
+        else:
+            self.selected_grad_image = new_gradient
 
         image_modify_object = ImgModifier(
-            outer_img=Image.open(os.path.join(gradient_path)),
+            outer_img=Image.open(os.path.join(self.selected_grad_image)),
             inner_img=Image.open(os.path.join(IMAGE_DIR_PATH, capture_name)),
             padding=padding
         )
@@ -143,11 +149,19 @@ class App(customtkinter.CTk):
         self.main_image.grid(row=0, column=0, padx=pad_x, pady=pad_y)
 
         # refresh gradient, generate random gradient
+        button_frame = customtkinter.CTkFrame(self.home_frame)
         regen_grad = customtkinter.CTkButton(
-            self.home_frame,
+            button_frame,
             text="Refresh Gradient", command=self.generate_gradient, border_spacing=10)
-        regen_grad.grid(row=2, column=0)
+        regen_grad.grid(row=0, column=0)
+        button_frame.grid(row=2, column=0)
 
+        # save modified image
+        save_button = customtkinter.CTkButton(
+            button_frame, text="Save Image", command=self.save_image, border_spacing=10)
+        save_button.grid(row=0, column=1, padx=20)
+
+        # checkbox for same color
         if self.use_same_colors is None:
             self.use_same_colors = customtkinter.CTkCheckBox(
                 self.home_frame, text="Keep Same Colors"
@@ -172,13 +186,27 @@ class App(customtkinter.CTk):
         slider.set(self.padding)
         slider_label = customtkinter.CTkLabel(sl_frame, text=f"Padding : {padding}x")
         slider_label.grid(row=0, column=0)
-        slider.grid(row=0, column=1)
+        slider.grid(row=0, column=1, padx=10)
+
         sl_frame.grid(row=4, column=0)
 
-        # save modified image
-        save_button = customtkinter.CTkButton(
-            self.home_frame, text="Save Image", command=self.save_image, border_spacing=10)
-        save_button.grid(row=5, column=0)
+        # scrolls gradients
+        if self.scrolls is None or not use_old:
+            self.scrolls = ScrollableLabelButtonFrame(
+                master=self.home_frame, width=200, height=150,
+                command=self.action_taken, corner_radius=10, orientation="horizontal"
+            )
+            self.scrolls.grid(row=5, column=0, padx=100, pady=10, sticky="nsew")
+
+            grad_images_dir = r"ui/premade_gradients"
+            gradients = [grad_images_dir + "/" + i for i in os.listdir(grad_images_dir)]
+
+            for i in range(len(gradients)):
+                self.scrolls.add_item(
+                    gradients[i], image=customtkinter.CTkImage(
+                        Image.open(gradients[i]), size=(200, 100)
+                    )
+                )
 
         self.deiconify()
 
@@ -220,7 +248,7 @@ class App(customtkinter.CTk):
         """
         Slider padding
         """
-        self.capture_screenshot(use_old=True, padding=int(slide))
+        self.capture_screenshot(use_old=True, padding=int(slide), new_gradient=self.selected_grad_image)
 
     def save_image(self):
         """save image to selected file"""
@@ -235,6 +263,12 @@ class App(customtkinter.CTk):
         if messagebox.askokcancel("Quit", "Do you want to quit, have you saved the image?"):
             tempdir.cleanup()
             self.destroy()
+
+    def action_taken(self, image):
+        """Action after click"""
+        self.capture_screenshot(
+            use_old=True, new_gradient=image
+        )
 
 
 if __name__ == "__main__":